+++ requake/scripts/requake.py
@@ -19,6 +19,9 @@ def main():
     if config.args.action == 'scan_catalog':
         from ..scan_catalog import scan_catalog
         scan_catalog(config)
+    if config.args.action == 'scan_templates':
+        from ..scan_templates import scan_templates
+        scan_templates(config)
     if config.args.action == 'plot_pair':
         from ..plot_pair import plot_pair
         plot_pair(config)